+++ NRG-APP/src/main/java/com/nrg/services/VehicleServiceImpl.java
@@ -5,7 +5,6 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.nrg.models.User;
 import com.nrg.models.VehicleDetails;
 import com.nrg.repositories.VehicleRepository;
 
@@ -16,7 +15,7 @@ public class VehicleServiceImpl implements VehicleService {
 	VehicleRepository vehicleRepository;
 	
 	@Override
-	public List<VehicleDetails> getVehiclesOfUser(Long userId) {
+	public List<VehicleDetails> getVehiclesOfUser(Integer userId) {
 		return vehicleRepository.findByUserId(userId);
 	}
